[refactor] 모임 상세 페이지 UI 컴포넌트 분리
