+++ src/components/gatherings/detail/loading/ThumbnailLoading.tsx
